Feat: 재료 즐겨찾기 등록
1. 재료 즐겨찾기 등록

Resolves: #03

+++ refrigerator/src/main/java/moja/refrigerator/dto/ingredient/request/RequestRegistIngredientBookmark.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class RequestRegistIngredientBookmark {
-    private int userPk;
-    private int ingredientPk;
+    private long userPk;
+    private long ingredientPk;
 }
